Refactor
Separate the Buffer class into its own file and move the search logic
into its own class. This decoupling will (hopefully) make it much easier
to change what happens with the search results.
